+++ apps/doc/src/lib/types/navs.ts
@@ -1,8 +1,9 @@
 export type NavItem = {
 	slug: string
 	title: string
-	label?: string
+	label: string
 	asset?: string
+	assetType?: string
 	reveal?: string
 	formaction?: string
 	actionPath?: string