+++ apps/doc/src/lib/ui/Socials.svelte
@@ -3,13 +3,14 @@
 		color?: string
 		variant?: string
 		links: {
-			title: string
+			label: string
 			url: string
 			shape?: string
 			color?: string
 			variant?: string
 			size: string
 			asset: string
+			assetType: string
 		}[]
 	}
 
@@ -17,13 +18,13 @@
 </script>
 
 <ul class="l:flex end unstyled">
-	{#each links as { title, url, shape, size, asset }, i (i)}
+	{#each links as { label, url, shape, size, asset, assetType }, i (i)}
 		<li>
 			<a
-				class={`${variant} shape:${shape} color:${color} ${asset} size:${size}`}
+				class={`${variant} shape:${shape} color:${color} ${assetType}:${asset} size:${size}`}
 				href={url}
-				{title}
-				aria-label={title}
+				title={label}
+				aria-label={label}
 			>
 			</a>
 		</li>