+++ packages/playbook/src/lib/fixtures/js/blocks.ts
@@ -1,7 +1,8 @@
 export default {
-	Button: {label: 'Button', asset: 'idea'},
+	Button: {label: 'Button', asset: 'idea', justify: 'start'},
 	Expand: {
 		title: 'Expand',
+		justify: 'start',
 		states: {
 			expanded: {
 				id: 'button-expand',
@@ -20,6 +21,7 @@ export default {
 	Switch: {
 		title: 'Switch',
 		shape: 'round',
+		justify: 'start',
 		states: {
 			active: {
 				id: 'button-switch',
@@ -35,8 +37,44 @@ export default {
 			},
 		},
 	},
+	Toggle: {
+		title: 'Toggle',
+		shape: 'round',
+		justify: 'start',
+		states: {
+			active: {
+				id: 'button-toggle',
+				label: 'Toggle In',
+				event: 'toggle',
+				asset: 'toggle-active',
+			},
+			inactive: {
+				id: 'button-toggle',
+				label: 'Toggle Out',
+				event: 'toggle',
+				asset: 'toggle-inactive',
+			},
+		},
+	},
 	InputCheck: {label: 'A checkbox choice'},
 	InputRadio: {label: 'A radio choice'},
+	InputGroup: {
+		legend: 'InputGroup',
+		// type: 'radio',
+		type: 'checkbox',
+		id: 'input-group',
+		name: 'input-group',
+		value: 'input-group',
+		variant: 'outline',
+		items: [
+			{label: 'Checkbox 1', id: 'input-group.check-1', value: 'check-1'},
+			{label: 'Checkbox 2', id: 'input-group.check-2', value: 'check-2'},
+			{label: 'Checkbox 3', id: 'input-group.check-3', value: 'check-3'},
+			{label: 'Checkbox 4', id: 'input-group.check-4', value: 'check-4'},
+			{label: 'Checkbox 5', id: 'input-group.check-5', value: 'check-5'},
+			{label: 'Checkbox 6', id: 'input-group.check-6', value: 'check-6'},
+		],
+	},
 	InputRange: {label: 'A range of ordered choices'},
 	Feedback: {
 		status: [
@@ -72,7 +110,6 @@ export default {
 			},
 		],
 	},
-	Toggle: {label: 'Toggle', asset: 'favorite'},
 	Fieldset: {
 		items: ['Form input 1', 'Form input 2', 'Form input 3'],
 	},