+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fat-fuzzy/config
 
+## 0.1.20
+
+### Patch Changes
+
+- 53ea789: InputGroup & Editor improvements, style fixes
+
 ## 0.1.19
 
 ### Patch Changes